% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_call_function_ext}
\alias{wasmer_call_function_ext}
\title{Call an exported function from a WASM instance.}
\usage{
wasmer_call_function_ext(ptr, instance_name, function_name, args)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{instance_name}{Name of the instance}

\item{function_name}{Name of the function to call}

\item{args}{Arguments as R list}
}
\value{
List with success flag and result or error
}
\description{
Call an exported function from a WASM instance.
}
