% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_compile_wat_ext}
\alias{wasmer_compile_wat_ext}
\title{Compile a WAT (WebAssembly Text) module and add it to the runtime.}
\usage{
wasmer_compile_wat_ext(ptr, wat_code, module_name)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{wat_code}{WAT code as a string}

\item{module_name}{Name to register the module under}
}
\value{
Status message
}
\description{
Compile a WAT (WebAssembly Text) module and add it to the runtime.
}
