% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_function_new_f64_f64_to_f64}
\alias{wasmer_function_new_f64_f64_to_f64}
\title{Create a Wasmer host function from an R function with static signature (f64, f64) -> f64}
\usage{
wasmer_function_new_f64_f64_to_f64(ptr, rfun)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{rfun}{R function object}
}
\value{
External pointer to Function
}
\description{
Create a Wasmer host function from an R function with static signature (f64, f64) -> f64
}
