% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_function_new_i32_to_i32}
\alias{wasmer_function_new_i32_to_i32}
\title{Create host function (i32 -> i32)}
\usage{
wasmer_function_new_i32_to_i32(ptr, rfun)
}
\description{
Create a WASM host function that takes an i32 and returns an i32, using an R function as the implementation.
}
\details{
Create a WASM host function with signature i32 -> i32
}
\examples{
wasmer_function_new_i32_to_i32(ptr, function(x) x)
}
\seealso{
\code{\link[=wasmer_register_r_function_ext]{wasmer_register_r_function_ext()}}, \code{\link[=wasmer_function_new_ext]{wasmer_function_new_ext()}}, \code{\link[=wasmer_function_new_i32_i32_to_i32]{wasmer_function_new_i32_i32_to_i32()}}, \code{\link[=wasmer_function_new_f64_f64_to_f64]{wasmer_function_new_f64_f64_to_f64()}}, \code{\link[=wasmer_function_new_f64_to_f64]{wasmer_function_new_f64_to_f64()}}, \code{\link[=wasmer_function_new_i32_to_void]{wasmer_function_new_i32_to_void()}}, \code{\link[=wasmer_function_new_void_to_i32]{wasmer_function_new_void_to_i32()}}

Other host function registration: 
\code{\link{wasmer_function_new_ext}()},
\code{\link{wasmer_function_new_f64_f64_to_f64}()},
\code{\link{wasmer_function_new_f64_to_f64}()},
\code{\link{wasmer_function_new_i32_i32_to_i32}()},
\code{\link{wasmer_function_new_i32_to_void}()},
\code{\link{wasmer_function_new_void_to_i32}()},
\code{\link{wasmer_register_r_function_ext}()}
}
\concept{host function registration}
