% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_function_new_static_ext}
\alias{wasmer_function_new_static_ext}
\title{Create a Wasmer host function from an R function with static signature (i32, i32) -> i32
This is required for WASM tables and funcref use.}
\usage{
wasmer_function_new_static_ext(ptr, rfun, `_name`)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{rfun}{R function object}

\item{_name}{Character string for registry name}
}
\value{
External pointer to Function
}
\description{
Create a Wasmer host function from an R function with static signature (i32, i32) -> i32
This is required for WASM tables and funcref use.
}
