% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_hello_world_example_ext}
\alias{wasmer_hello_world_example_ext}
\title{Simple Hello World WASM example}
\usage{
wasmer_hello_world_example_ext(runtime)
}
\arguments{
\item{runtime}{External pointer to WasmerRuntime}
}
\value{
String result from WASM hello function
}
\description{
Simple Hello World WASM example
}
