% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_instantiate_ext}
\alias{wasmer_instantiate_ext}
\title{Instantiate a compiled module in the runtime.}
\usage{
wasmer_instantiate_ext(ptr, module_name, instance_name)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{module_name}{Name of the module to instantiate}

\item{instance_name}{Name to register the instance under}
}
\value{
Status message
}
\description{
Instantiate a compiled module in the runtime.
}
