% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_instantiate_with_math_imports_ext}
\alias{wasmer_instantiate_with_math_imports_ext}
\title{Create an instance with host functions for mathematical operations}
\usage{
wasmer_instantiate_with_math_imports_ext(ptr, module_name, instance_name)
}
\arguments{
\item{module_name}{String name of the module to instantiate}

\item{instance_name}{String name to identify this instance}

\item{runtime}{External pointer to WasmerRuntime}
}
\value{
Status message
}
\description{
Create an instance with host functions for mathematical operations
}
