% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_memory_read_string_ext}
\alias{wasmer_memory_read_string_ext}
\title{Read UTF-8 string from WASM memory}
\usage{
wasmer_memory_read_string_ext(ptr, instance_name, memory_name, offset, length)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{instance_name}{Name of the instance}

\item{memory_name}{Name of the exported memory}

\item{offset}{Offset to start reading}

\item{length}{Number of bytes to read}
}
\value{
String
}
\description{
Read UTF-8 string from WASM memory
}
