% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_memory_size_ext}
\alias{wasmer_memory_size_ext}
\title{Get the size of exported memory (in bytes and pages)}
\usage{
wasmer_memory_size_ext(ptr, instance_name, memory_name)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{instance_name}{Name of the instance}

\item{memory_name}{Name of the exported memory (default "memory")}
}
\value{
List with size_bytes and size_pages
}
\description{
Get the size of exported memory (in bytes and pages)
}
