% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_memory_write_ext}
\alias{wasmer_memory_write_ext}
\title{Write bytes to WASM memory}
\usage{
wasmer_memory_write_ext(ptr, instance_name, memory_name, offset, bytes)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{instance_name}{Name of the instance}

\item{memory_name}{Name of the exported memory}

\item{offset}{Offset to start writing}

\item{bytes}{Raw vector of bytes to write}
}
\value{
TRUE if successful
}
\description{
Write bytes to WASM memory
}
