% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_register_r_function_ext}
\alias{wasmer_register_r_function_ext}
\title{Register an R function for use as a host function in WASM.}
\usage{
wasmer_register_r_function_ext(ptr, name, fun)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{name}{Name to register the function under}

\item{fun}{R function object}
}
\value{
TRUE if successful
}
\description{
Register an R function for use as a host function in WASM.
}
