% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_runtime_new}
\alias{wasmer_runtime_new}
\title{Create a new Wasmer runtime for R. Returns an external pointer to the runtime object.}
\usage{
wasmer_runtime_new()
}
\value{
External pointer to WasmerRuntime
}
\description{
Create a new Wasmer runtime for R. Returns an external pointer to the runtime object.
}
