% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_runtime_release_ressources}
\alias{wasmer_runtime_release_ressources}
\title{Explicitly shutdown the runtime, free resources, and clear the R external pointer}
\usage{
wasmer_runtime_release_ressources(ptr)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}
}
\description{
Explicitly shutdown the runtime, free resources, and clear the R external pointer
}
