% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_table_get_ext}
\alias{wasmer_table_get_ext}
\title{Get a function reference from a WASM Table}
\usage{
wasmer_table_get_ext(ptr, table_ptr, index)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{table_ptr}{External pointer to Table}

\item{index}{Index to get}
}
\value{
External pointer to Function (or NULL)
}
\description{
Get a function reference from a WASM Table
}
