% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_table_grow_ext}
\alias{wasmer_table_grow_ext}
\title{Grow a WASM Table}
\usage{
wasmer_table_grow_ext(ptr, table_ptr, delta, func_ptr)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{table_ptr}{External pointer to Table}

\item{delta}{Number of elements to grow}

\item{func_ptr}{External pointer to Function to fill new slots}
}
\value{
Previous size
}
\description{
Grow a WASM Table
}
