% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_table_new_ext}
\alias{wasmer_table_new_ext}
\title{Create a new WASM Table}
\usage{
wasmer_table_new_ext(ptr, min, max)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{min}{Minimum size}

\item{max}{Maximum size (optional)}
}
\value{
External pointer to Table
}
\description{
Create a new WASM Table
}
