% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_table_set_ext}
\alias{wasmer_table_set_ext}
\title{Set a function reference in a WASM Table}
\usage{
wasmer_table_set_ext(ptr, table_ptr, index, func_ptr)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{table_ptr}{External pointer to Table}

\item{index}{Index to set}

\item{func_ptr}{External pointer to Function}
}
\value{
TRUE if successful
}
\description{
Set a function reference in a WASM Table
}
