% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_wat_to_wasm_ext}
\alias{wasmer_wat_to_wasm_ext}
\title{Convert WAT (WebAssembly Text) to WASM binary and return as R raw vector}
\usage{
wasmer_wat_to_wasm_ext(wat_code)
}
\arguments{
\item{wat_code}{WAT code as a string}
}
\value{
WASM binary as R raw vector, or error string if conversion fails
}
\description{
Convert WAT (WebAssembly Text) to WASM binary and return as R raw vector
}
