% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_compile_wasm_ext}
\alias{wasmer_compile_wasm_ext}
\title{Compile a WASM binary and add it to the runtime.}
\usage{
wasmer_compile_wasm_ext(ptr, wasm_bytes, module_name)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{wasm_bytes}{WASM binary as R raw vector}

\item{module_name}{Name to register the module under}
}
\value{
Status message
}
\description{
Compile a WASM binary and add it to the runtime.
}
