% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_function_new_ext}
\alias{wasmer_function_new_ext}
\title{Create a Wasmer host function from an R function with dynamic signature}
\usage{
wasmer_function_new_ext(ptr, rfun, arg_types, ret_types, `_name`)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{rfun}{R function object}

\item{arg_types}{Character vector of argument types (e.g. c("i32", "f64"))}

\item{ret_types}{Character vector of return types (e.g. c("i32"))}

\item{name}{Character string for registry name}
}
\value{
External pointer to Function
}
\description{
Create a Wasmer host function from an R function with dynamic signature
}
