% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_function_new_i32_to_void}
\alias{wasmer_function_new_i32_to_void}
\title{Create a Wasmer host function from an R function with static signature (i32) -> void}
\usage{
wasmer_function_new_i32_to_void(ptr, rfun)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{rfun}{R function object}
}
\value{
External pointer to Function
}
\description{
Create a Wasmer host function from an R function with static signature (i32) -> void
}
