% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_get_exported_table_ext}
\alias{wasmer_get_exported_table_ext}
\title{Get a pointer to an exported table from a WASM instance by name.}
\usage{
wasmer_get_exported_table_ext(ptr, instance_name, table_export_name)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{instance_name}{Name of the instance}

\item{table_export_name}{Name of the exported table}
}
\value{
External pointer to Table, or NULL if not found
}
\description{
Get a pointer to an exported table from a WASM instance by name.
}
