% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_hello_world_example_ext}
\alias{wasmer_hello_world_example_ext}
\title{Create a simple "Hello World" example}
\usage{
wasmer_hello_world_example_ext(ptr)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}
}
\value{
String result from WASM hello function
}
\description{
Create a simple "Hello World" example
}
