% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_host_function_example_ext}
\alias{wasmer_host_function_example_ext}
\title{Example with host function imports}
\usage{
wasmer_host_function_example_ext(ptr)
}
\arguments{
\item{runtime}{External pointer to WasmerRuntime}
}
\value{
List with results
}
\description{
Example with host function imports
}
