% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_instantiate_with_table_ext}
\alias{wasmer_instantiate_with_table_ext}
\title{Instantiate a compiled module in the runtime, with a custom table import.}
\usage{
wasmer_instantiate_with_table_ext(ptr, module_name, instance_name, table_ptr)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{module_name}{Name of the module to instantiate}

\item{instance_name}{Name to register the instance under}

\item{table_ptr}{External pointer to Table to import as "env.host_table"}
}
\value{
Status message
}
\description{
Instantiate a compiled module in the runtime, with a custom table import.
}
