% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_list_exports_ext}
\alias{wasmer_list_exports_ext}
\title{List all exports from a WASM instance.}
\usage{
wasmer_list_exports_ext(ptr, instance_name)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{instance_name}{Name of the instance}
}
\value{
List with success flag and exports or error
}
\description{
List all exports from a WASM instance.
}
