% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_list_function_signatures_ext}
\alias{wasmer_list_function_signatures_ext}
\title{List exported function signatures (name, input types, output types) for a WASM instance}
\usage{
wasmer_list_function_signatures_ext(ptr, instance_name)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{instance_name}{Name of the instance}
}
\value{
Data frame with columns: name, params, results
}
\description{
List exported function signatures (name, input types, output types) for a WASM instance
}
