% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_math_example_ext}
\alias{wasmer_math_example_ext}
\title{Math operations example}
\usage{
wasmer_math_example_ext(runtime, a, b)
}
\arguments{
\item{runtime}{External pointer to WasmerRuntime}

\item{a}{First integer}

\item{b}{Second integer}
}
\value{
List with results of add and multiply
}
\description{
Math operations example
}
