% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_memory_write_ext}
\alias{wasmer_memory_write_ext}
\title{Write WASM memory}
\usage{
wasmer_memory_write_ext(ptr, instance_name, memory_name, offset, bytes)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime.}

\item{instance_name}{Name of the instance.}

\item{memory_name}{Name of the exported memory.}

\item{offset}{Offset to start writing.}

\item{bytes}{Raw vector of bytes to write.}
}
\value{
TRUE if successful
}
\description{
Write bytes to WASM memory.
}
\details{
Write bytes to WASM memory
}
\examples{
wasmer_memory_write_ext(ptr, "inst1", "memory", 0, as.raw(c(1,2,3)))
}
\seealso{
\code{\link[=wasmer_memory_size_ext]{wasmer_memory_size_ext()}}, \code{\link[=wasmer_memory_read_ext]{wasmer_memory_read_ext()}}, \code{\link[=wasmer_memory_read_string_ext]{wasmer_memory_read_string_ext()}}, \code{\link[=wasmer_memory_grow_ext]{wasmer_memory_grow_ext()}}

Other memory operations: 
\code{\link{wasmer_memory_grow_ext}()},
\code{\link{wasmer_memory_read_ext}()},
\code{\link{wasmer_memory_read_string_ext}()},
\code{\link{wasmer_memory_size_ext}()}
}
\concept{memory operations}
