% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_runtime_new}
\alias{wasmer_runtime_new}
\title{Create a new Wasmer runtime}
\usage{
wasmer_runtime_new()
}
\value{
External pointer to WasmerRuntime
}
\description{
Create a new Wasmer runtime for executing WebAssembly modules.
}
\details{
Create a new Wasmer runtime
}
\examples{
\dontrun{
ptr <- wasmer_runtime_new()
}
}
\seealso{
\code{\link[=wasmer_runtime_new_with_compiler_ext]{wasmer_runtime_new_with_compiler_ext()}}, \code{\link[=wasmer_runtime_release_ressources]{wasmer_runtime_release_ressources()}}

Other runtime management: 
\code{\link{wasmer_runtime_new_with_compiler_ext}()},
\code{\link{wasmer_runtime_release_ressources}()}
}
\concept{runtime management}
