% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_runtime_new_with_compiler_ext}
\alias{wasmer_runtime_new_with_compiler_ext}
\title{Create a new Wasmer runtime with a specific compiler}
\usage{
wasmer_runtime_new_with_compiler_ext(compiler_name)
}
\arguments{
\item{compiler_name}{Name of the compiler ("cranelift", "singlepass").}
}
\value{
External pointer to WasmerRuntime
}
\description{
Create a new Wasmer runtime for executing WebAssembly modules using a specified compiler backend.
}
\details{
Create a new Wasmer runtime with a specific compiler
}
\examples{
\dontrun{
ptr <- wasmer_runtime_new_with_compiler_ext("cranelift")
}
}
\seealso{
\code{\link[=wasmer_runtime_new]{wasmer_runtime_new()}}, \code{\link[=wasmer_runtime_release_ressources]{wasmer_runtime_release_ressources()}}

Other runtime management: 
\code{\link{wasmer_runtime_new}()},
\code{\link{wasmer_runtime_release_ressources}()}
}
\concept{runtime management}
