% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_table_new_ext}
\alias{wasmer_table_new_ext}
\title{Create WASM Table}
\usage{
wasmer_table_new_ext(ptr, min, max)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime.}

\item{min}{Minimum size.}

\item{max}{Maximum size (optional).}
}
\value{
External pointer to Table
}
\description{
Create a new WASM Table.
}
\details{
Create a new WASM Table
}
\examples{
\dontrun{
wasmer_table_new_ext(ptr, 1, 10)
}
}
\seealso{
\code{\link[=wasmer_table_set_ext]{wasmer_table_set_ext()}}, \code{\link[=wasmer_table_grow_ext]{wasmer_table_grow_ext()}}, \code{\link[=wasmer_table_get_ext]{wasmer_table_get_ext()}}, \code{\link[=wasmer_get_exported_table_ext]{wasmer_get_exported_table_ext()}}

Other table operations: 
\code{\link{wasmer_get_exported_table_ext}()},
\code{\link{wasmer_table_get_ext}()},
\code{\link{wasmer_table_grow_ext}()},
\code{\link{wasmer_table_set_ext}()}
}
\concept{table operations}
