% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_table_set_ext}
\alias{wasmer_table_set_ext}
\title{Set WASM Table entry}
\usage{
wasmer_table_set_ext(ptr, table_ptr, index, func_ptr)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime.}

\item{table_ptr}{External pointer to Table.}

\item{index}{Index to set.}

\item{func_ptr}{External pointer to Function.}
}
\value{
TRUE if successful
}
\description{
Set a function reference in a WASM Table.
}
\details{
Set a function reference in a WASM Table
}
\examples{
wasmer_table_set_ext(ptr, table_ptr, 0, func_ptr)
}
\seealso{
\code{\link[=wasmer_table_new_ext]{wasmer_table_new_ext()}}, \code{\link[=wasmer_table_grow_ext]{wasmer_table_grow_ext()}}, \code{\link[=wasmer_table_get_ext]{wasmer_table_get_ext()}}, \code{\link[=wasmer_get_exported_table_ext]{wasmer_get_exported_table_ext()}}

Other table operations: 
\code{\link{wasmer_get_exported_table_ext}()},
\code{\link{wasmer_table_get_ext}()},
\code{\link{wasmer_table_grow_ext}()},
\code{\link{wasmer_table_new_ext}()}
}
\concept{table operations}
