% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wasmer_wasi_state_new_ext}
\alias{wasmer_wasi_state_new_ext}
\title{Create a WASI state for the runtime.}
\usage{
wasmer_wasi_state_new_ext(ptr, module_name)
}
\arguments{
\item{ptr}{External pointer to WasmerRuntime}

\item{module_name}{Name of the module (for WASI args)}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Create a WASI state for the runtime.
}
